requirements.txt
python-dotenv==1.0.1
requests==2.32.3
python-slugify==8.0.1
pygithub==2.3.0